mpegmap/HTMLSourcingInbandTracks-3.4-1.html

<!doctype html>
<html>
  <head>
    <title>TC 5.1.2-1 VideoTrack</title>
    <link rel="author" title="CableLabs" href="html5@cablelabs.com">
    <meta name="flags" content="[requirement flags]">
    <meta name="assert" content="MPEG-2 SPTS - UA creates VideoTrack, attributes set.">
    <meta name="timeout" content="long">
    <script src="testmedia.js"></script>
    <!-- Source HTTP server local copy of World Wide Web Consortium (W3C) Test Harness. --> 
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <!-- Script Version: BETA -->
    <!-- About the Media Resource

         To verify DOM elements the details of the test media must be known prior to test execution.
         This test script is bound to the test media file sourced in the script.
         Sourcing a differnt test media file, or modifying that file, may invalidate the test.

         The media is an MPEG-2, Single Program Transport Stream
         containing one video elementary stream, ES PID 33.
    -->
  </head>
  <body>
    <p>
      <a href="http://www.w3.org/TR/html5/references.html#refsINBANDTRACKS">Spec
      Reference [INBANDTRACKS]</a>
    </p>
    <video id="v" width="400" height="300" autoplay controls></video>
    <div id="log"></div>
    <script>
      var test = async_test("VideoTrack attributes.");
      var vid = document.getElementById("v");
      vid.src = mediaServerURL() + "UserPrivateStreams.ts";

      var pid = "33"; // video ES PID
      var kind = "main";

      vid.videoTracks.addEventListener("addtrack", function(event) {
        var track = event.track;
        test.step(function() {
          assert_equals(vid.videoTracks.length, 1, "VideoTracks.length");
          assert_equals(track.id, pid, "VideoTrack.id");
          assert_equals(track.kind, kind, "VideoTrack.kind");
        });
        test.done();
      });

      vid.addEventListener("playing", function() {
        vid.pause();
        if (!vid.videoTracks.getTrackById(pid)) {
          test.step(function() {
            assert_unreached("Did not find VideoTrack PID " + pid + " in DOM");
          }, "Did not find VideoTrack " + pid + " in DOM");
          test.done();
        }
      });
    </script>
  </body>
</html>
